back/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
line-bot-sdk
requests
selenium
pandas
tqdm
BeautifulSoup4
requests
lxml